Match fdlibm & PPCArch.c
fdlibm seems to be available in archives and already used in pikmin, so
I've just grabbed the files from it used by sms. Note that fabs mangling
seems to be screwed up in fdlibm TUs somehow and is c++-style even
though the library itself is C, so for now we work around thata instead
of trying to figure out what originally happened.

Also cleaned up PPCArch.c to be at 100%.